memory tiering: check order in migrate_balanced_pgdat()
When THP is promoted from the slow memory node to the fast memory
node, migrate_balanced_pgdat() is called to check whether there's
enough free space on the target node.  But originally, only the total
number of the free pages is considered, but the order of the free
pages isn't considered.  So it's possible that there's no free page
with expected order, although the total number of the free pages is
enough.  So that the page demotion for the fast memory node isn't
initiated to make the high order pages available.

To solve the above issue, the page order is considered in
migrate_balanced_pgdat().

TODO: test results?

Signed-off-by: "Huang, Ying" <ying.huang@intel.com>